+++ src/components/LastInteractions.js
@@ -75,7 +75,7 @@ const LastInteractions = () => {
   }, [activity.categories, safe.currentAccount]);
 
   return (
-    <Grid alignItems="center" container justify="center" spacing={2}>
+    <Grid alignItems="center" container justifyContent="center" spacing={2}>
       {lastActiveProfiles.map(({ safeAddress }) => {
         return (
           <LastInteractionsAccount address={safeAddress} key={safeAddress} />